# Experimental SIMD-accelerated parsing using AVX-512