--- a/src/Engine/ByteBlockBackedDictionary.cpp
+++ b/src/Engine/ByteBlockBackedDictionary.cpp
@@ -21,6 +21,17 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 // OTHER DEALINGS IN THE SOFTWARE.
 
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+#if defined(__AVX512F__)
+#include <immintrin.h>
+
+#include <cstdint>
+#include <memory>
+#else
+#error AVX512 support required
+#endif
+#endif
+
 #include "ByteBlockBackedDictionary.h"
 
 namespace McBopomofo {
@@ -74,6 +85,7 @@ const char* AdvanceToNextNonContentCharacter(const char* ptr, const char* end) {
   return ptr;
 }
 
+#ifndef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
 const char* FindFirstNULL(const char* ptr, const char* end,
                           size_t* firstLineNumber = nullptr) {
   const char* i = ptr;
@@ -98,10 +110,168 @@ const char* FindFirstNULL(const char* ptr, const char* end,
 
   return i;
 }
+#endif
 
 bool IsCRLF(char c) { return c == '\n' || c == '\r'; }
 
 bool IsWhitespace(char c) { return c == ' ' || c == '\t'; }
+
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+
+const char* AVX512_AdvanceToNextCRLF(const char* ptr,
+                                     const char* unaligned32End,
+                                     const char* end) {
+  const __m256i lfs = _mm256_set1_epi8('\n');
+  const __m256i crs = _mm256_set1_epi8('\r');
+
+  while (ptr < unaligned32End) {
+    const __m256i block = _mm256_loadu_epi8(ptr);
+    const __mmask32 foundLFs = _mm256_cmpeq_epi8_mask(block, lfs);
+    const __mmask32 foundCRs = _mm256_cmpeq_epi8_mask(block, crs);
+    const __mmask32 mask = _kor_mask32(foundLFs, foundCRs);
+    if (mask != 0) {
+      return ptr + _tzcnt_u64(mask);
+    }
+
+    ptr += 32;
+  }
+
+  return AdvanceToNextCRLF(ptr, end);
+}
+
+// Four chars: 0x09 (T), 0x0a (L), 0x0d (C), 0x20 (S)
+// T maps to 0x01
+// L maps to 0x02
+// C maps to 0x04
+// T|L|C = 0x07
+// S maps to 0x08
+alignas(32) constexpr uint8_t LO_NIBBLES_LOOKUP[32] = {
+    0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x02,
+    0x00, 0x00, 0x04, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x01, 0x02, 0x00, 0x00, 0x04, 0x00, 0x00,
+};
+
+alignas(32) constexpr uint8_t HI_NIBBLES_LOOKUP[32] = {
+    0x07, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x08, 0x00, 0x00, 0x00,
+    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+};
+
+const char* AVX512_AdvanceToNextNonContentCharacter(const char* ptr,
+                                                    const char* unaligned32End,
+                                                    const char* end) {
+  while (ptr < unaligned32End) {
+    const __m256i input = _mm256_loadu_epi8(ptr);
+
+    const __m256i mask = _mm256_set1_epi8(0x0f);
+    const __m256i loNibbles = _mm256_and_si256(input, mask);
+    const __m256i hiNibbles =
+        _mm256_and_si256(_mm256_srli_epi16(input, 4), mask);
+    const __m256i loTbl =
+        _mm256_load_si256(reinterpret_cast<const __m256i*>(LO_NIBBLES_LOOKUP));
+    const __m256i lo = _mm256_shuffle_epi8(loTbl, loNibbles);
+    const __m256i hiTbl =
+        _mm256_load_si256(reinterpret_cast<const __m256i*>(HI_NIBBLES_LOOKUP));
+    const __m256i hi = _mm256_shuffle_epi8(hiTbl, hiNibbles);
+    const __m256i intersection = _mm256_and_si256(lo, hi);
+    const __mmask32 nonContentMask =
+        _mm256_cmpneq_epi8_mask(intersection, _mm256_setzero_si256());
+    if (nonContentMask != 0) {
+      int64_t firstNonContentIndex = _tzcnt_u64(nonContentMask);
+      return ptr + firstNonContentIndex;
+    }
+    ptr += 32;
+  }
+
+  return AdvanceToNextNonContentCharacter(ptr, end);
+}
+
+constexpr uintptr_t ALIGN64 = 64;
+constexpr uintptr_t ALIGN64_MASK = ALIGN64 - 1;
+
+const char* AVX512_FindFirstNULL(const char* ptr, const char* end,
+                                 size_t* firstLineNumber = nullptr) {
+  const char* i = ptr;
+  bool found = false;
+  if ((reinterpret_cast<uintptr_t>(i) & ALIGN64_MASK) != 0) {
+    const char* headEnd = reinterpret_cast<const char*>(
+        reinterpret_cast<uintptr_t>(i + ALIGN64_MASK) & ~ALIGN64_MASK);
+    headEnd = headEnd < end ? headEnd : end;
+    while (i != headEnd) {
+      if (*i == '\0') {
+        found = true;
+        break;
+      }
+      ++i;
+    }
+  }
+
+  if (!found && i != end) {
+    const char* middleEnd = reinterpret_cast<const char*>(
+        reinterpret_cast<uintptr_t>(end) & ~ALIGN64_MASK);
+    const __m512i zeros = _mm512_setzero_si512();
+    while (i != middleEnd) {
+      const __m512i block = _mm512_load_si512(i);
+      const __mmask64 mask = _mm512_cmpeq_epi8_mask(block, zeros);
+      if (mask != 0) {
+        found = true;
+        i += _tzcnt_u64(mask);
+        break;
+      }
+      i += ALIGN64;
+    }
+  }
+
+  if (!found) {
+    while (i != end) {
+      if (*i == 0) {
+        found = true;
+        break;
+      }
+      ++i;
+    }
+  }
+
+  if (!found || firstLineNumber == nullptr) {
+    return i;
+  }
+
+  size_t lineCounter = 1;
+  if ((reinterpret_cast<uintptr_t>(ptr) & ALIGN64_MASK) != 0) {
+    const char* headEnd = reinterpret_cast<const char*>(
+        reinterpret_cast<uintptr_t>(ptr + ALIGN64_MASK) & ~ALIGN64_MASK);
+    headEnd = headEnd < i ? headEnd : i;
+    while (ptr != headEnd) {
+      if (*ptr == '\n') {
+        ++lineCounter;
+      }
+      ++ptr;
+    }
+  }
+
+  if (ptr != i) {
+    const char* middleEnd = reinterpret_cast<const char*>(
+        reinterpret_cast<uintptr_t>(i) & ~ALIGN64_MASK);
+    const __m512i linefeeds = _mm512_set1_epi8('\n');
+    while (ptr != middleEnd) {
+      const __m512i block = _mm512_load_si512(ptr);
+      const __mmask64 mask = _mm512_cmpeq_epi8_mask(block, linefeeds);
+      lineCounter += _mm_popcnt_u64(mask);
+      ptr += ALIGN64;
+    }
+  }
+
+  while (ptr != i) {
+    if (*ptr == '\n') {
+      ++lineCounter;
+    }
+    ++ptr;
+  }
+  *firstLineNumber = lineCounter;
+  return i;
+}
+
+#endif
 
 }  // namespace
 
@@ -131,9 +301,22 @@ bool ByteBlockBackedDictionary::parse(const char* block, size_t size,
   const char* ptr = block;
   const char* end = ptr + size;
 
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+  const char* unaligned32End =
+      reinterpret_cast<const char*>(reinterpret_cast<uintptr_t>(end) - 32);
+  if (unaligned32End < ptr) {
+    unaligned32End = ptr;
+  }
+#endif
+
   // Validate that no NULL characters are in the text.
   size_t errorAtLine = 0;
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+  const char* ctrlCharPtr = AVX512_FindFirstNULL(ptr, end, &errorAtLine);
+#else
   const char* ctrlCharPtr = FindFirstNULL(ptr, end, &errorAtLine);
+#endif
+
   if (ctrlCharPtr != end) {
     issues_.emplace_back(Issue::Type::NULL_CHARACTER_IN_TEXT, errorAtLine);
     return false;
@@ -149,12 +332,20 @@ bool ByteBlockBackedDictionary::parse(const char* block, size_t size,
       }
 
       if (*ptr == '#') {
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+        ptr = AVX512_AdvanceToNextCRLF(ptr, unaligned32End, end);
+#else
         ptr = AdvanceToNextCRLF(ptr, end);
+#endif
         continue;
       }
 
       const char* keyStart = ptr;
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+      ptr = AVX512_AdvanceToNextNonContentCharacter(ptr, unaligned32End, end);
+#else
       ptr = AdvanceToNextNonContentCharacter(ptr, end);
+#endif
       const char* keyEnd = ptr;
 
       ptr = AdvanceToNextNonWhitespace(ptr, end);
@@ -167,7 +358,11 @@ bool ByteBlockBackedDictionary::parse(const char* block, size_t size,
       }
 
       const char* valueStart = ptr;
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+      ptr = AVX512_AdvanceToNextCRLF(ptr, unaligned32End, end);
+#else
       ptr = AdvanceToNextCRLF(ptr, end);
+#endif
       const char* valueEnd = ptr;
 
       if (valueEnd == valueStart) {
@@ -212,12 +407,20 @@ bool ByteBlockBackedDictionary::parse(const char* block, size_t size,
       }
 
       if (*ptr == '#') {
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+        ptr = AVX512_AdvanceToNextCRLF(ptr, unaligned32End, end);
+#else
         ptr = AdvanceToNextCRLF(ptr, end);
+#endif
         continue;
       }
 
       const char* valueStart = ptr;
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+      ptr = AVX512_AdvanceToNextNonContentCharacter(ptr, unaligned32End, end);
+#else
       ptr = AdvanceToNextNonContentCharacter(ptr, end);
+#endif
       const char* valueEnd = ptr;
 
       ptr = AdvanceToNextNonWhitespace(ptr, end);
@@ -229,7 +432,11 @@ bool ByteBlockBackedDictionary::parse(const char* block, size_t size,
       }
 
       const char* maybeKeyStart = ptr;
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+      ptr = AVX512_AdvanceToNextNonContentCharacter(ptr, unaligned32End, end);
+#else
       ptr = AdvanceToNextNonContentCharacter(ptr, end);
+#endif
       const char* maybeKeyEnd = ptr;
       if (maybeKeyStart == maybeKeyEnd) {
         if (issues_.size() < MAX_ISSUES) {
@@ -252,7 +459,11 @@ bool ByteBlockBackedDictionary::parse(const char* block, size_t size,
         // More content incoming.
         valueEnd = maybeKeyEnd;
         maybeKeyStart = ptr;
+#ifdef ENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512
+        ptr = AVX512_AdvanceToNextNonContentCharacter(ptr, unaligned32End, end);
+#else
         ptr = AdvanceToNextNonContentCharacter(ptr, end);
+#endif
         maybeKeyEnd = ptr;
       }