### **User description**
This PR further speeds up `ByteBlockBackedDictionary`'s parser for over 50% by taking advantage of SIMD in three operations. The new code can only be run on processors with AVX-512 support (which most Intel and AMD CPUs in the last 3 years), and currently this is an experimental feature that needs to be enabled by passing this to CMake:

```
cmake -B build \
  -DCMAKE_INSTALL_PREFIX=/usr
  -DCMAKE_BUILD_TYPE=Release
  -DENABLE_EXPERIMENTAL_SIMD_SUPPORT_AVX512=1
```

Three operations were identified as the low-hanging fruits that benefited from SIMD: advancing to the next CR/LF, advancing to the next non-content character (defined as any of `'\t'`, `'\n'`, `'\r'`, and `' '`), and identifing whether there is any NULL characters (`'\0'`) in the input text. Both CR/LF and non-content character advancements load 32 bytes per iteration, where as NULL-character detection loads 64 bytes per iteration. The numbers of bytes are chosen based on the nature of the work. For example, loading 16 bytes per CR/LF advancement iteration would under-utilize SIMD, since the user phrase key column (the last column) can often be longer than 16 bytes, but loading 64 bytes per iteration would result in a lot of wasted work.

The techniques, espeically the one that uses the `VPSHUFB` for vectorized lookups, are well-documented in sources like the famous [simdjson paper](https://arxiv.org/abs/1902.08318) and Wojciech Muła's [article](http://0x80.pl/notesen/2018-10-18-simd-byte-lookup.html).

There are many future works here: supporting different architectures (especially ARM64), putting the right bulid and code structure in place, and perhaps even a rethinking of the parsing strategy. Also note `ByteBlockBackedDictionary::parse()` now spends half of its execution time building the actual `std::unordered_map`, and so perhaps that's another venue to pursue.

I'd say our current non-SIMD parser is already a big improvement over the previous parser. It's likely most user-defined phrase data will never exceed a few thousand entries (so one or two MBs, max), and a non-SIMD parser just works fine with sub-millisecond parsing time. Not to mention that the current PR doesn't make the parser reach simdjson-level speed. Still, I think it's something worth exploring and experimenting with.

Benchmarks on a mid-range Intel laptop below. First, the plain, non-SIMD version:

```
------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations
------------------------------------------------------------------------------------------------
BM_ByteBlockBackedDictionaryParseTest                    4379076 ns      4378553 ns          162
BM_ByteBlockBackedDictionaryValueColumnFirstParseTest    4858093 ns      4856630 ns          145
BM_KeyValueBlobReaderParseTest                           9294404 ns      9291123 ns           75
```

With SIMD (the old parser's benchmark provided as a reference):

```
------------------------------------------------------------------------------------------------
Benchmark                                                      Time             CPU   Iterations
------------------------------------------------------------------------------------------------
BM_ByteBlockBackedDictionaryParseTest                    2816968 ns      2816391 ns          252
BM_ByteBlockBackedDictionaryValueColumnFirstParseTest    3156197 ns      3155729 ns          221
BM_KeyValueBlobReaderParseTest                           8794783 ns      8793732 ns           77
```

So the SIMD-accelerated parser is 52–55% faster than the non-SIMD version, and is easily 2.8x–3.3x faster than the previous parser.


___

### **PR Type**
Enhancement, Tests


___

### **Description**
- Add AVX-512 SIMD parsing fast paths

- Fallback to scalar when disabled

- CMake option to enable AVX-512 flags

- Add large NULL-detection tests


___

### Diagram Walkthrough


```mermaid
flowchart LR
  parser["ByteBlockBackedDictionary::parse()"] -- "SIMD enabled" --> simd["AVX-512 helpers (CR/LF, non-content, NULL)"]
  parser -- "SIMD disabled" --> scalar["Scalar helpers (existing)"]
  simd -- "vectorized search" --> results["Positions / validation"]
  scalar -- "loop search" --> results
  build["CMake AVX-512 flag"] -- "enables intrinsics" --> simd
  tests["New large NULL tests"] -- "validate behavior" --> parser
```



<details> <summary><h3> File Walkthrough</h3></summary>

<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>ByteBlockBackedDictionary.cpp</strong><dd><code>AVX-512 accelerated parsing with scalar fallback</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

src/Engine/ByteBlockBackedDictionary.cpp

<ul><li>Add AVX-512 intrinsics-based scanners<br> <li> Wire SIMD paths under compile-time flag<br> <li> Preserve scalar fallbacks for portability<br> <li> Optimize line counting with vector popcount</ul>


</details>


  </td>
  <td><a href="https://github.com/openvanilla/fcitx5-mcbopomofo/pull/194/files#diff-b15b618147c24bee4764701a322bd8cb58c85883d00331487a234496f1644281">+210/-0</a>&nbsp; </td>

</tr>
</table></td></tr><tr><td><strong>Tests</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>ByteBlockBackedDictionaryTest.cpp</strong><dd><code>Tests for large-scale NULL detection</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

src/Engine/ByteBlockBackedDictionaryTest.cpp

<ul><li>Add large-input NULL detection tests<br> <li> Verify issue type and line number<br> <li> Stress test mixed newlines and NULLs</ul>


</details>


  </td>
  <td><a href="https://github.com/openvanilla/fcitx5-mcbopomofo/pull/194/files#diff-e62855e535ad02e6bb2ae225f0a7bd4885e845fb1eb5b5e7e8e993029a5e2e76">+29/-0</a>&nbsp; &nbsp; </td>

</tr>
</table></td></tr><tr><td><strong>Configuration changes</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>CMakeLists.txt</strong><dd><code>CMake flag to enable AVX-512 SIMD build</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

src/Engine/CMakeLists.txt

<ul><li>Add AVX-512 compile options flag<br> <li> Minor formatting cleanup<br> <li> Define feature macro when enabled</ul>


</details>


  </td>
  <td><a href="https://github.com/openvanilla/fcitx5-mcbopomofo/pull/194/files#diff-6f9e82a90400b1432e336c62c68dfd8cd5ab0df6650d7e9606af68e241b3fefa">+7/-3</a>&nbsp; &nbsp; &nbsp; </td>

</tr>
</table></td></tr></tr></tbody></table>

</details>

___